Move cargo package dependency types to separate module
